face-mask-detector/requirements.txt
gradio==4.29.0
torch
torchvision
pillow
mediapipe
numpy
opencv-python